requirements.txt
weaviate-client==4.5.1
streamlit==1.32.0
tqdm==4.66.2
streamlit-image-select==0.6.0
